pyproject.toml
[tool.poetry]
name = "clean-tech-rag"
version = "0.1.0"
description = ""
authors = [ "Etienne Roulet <etienne.roulet@student.fhnw.ch>","Janis Frölich <janis.frölich@students.fhnw.ch>","Nicola Rohner <nicola.rohner@students.fhnw.ch>", "Patrick Schürmann <patrick.schürmann@students.fhnw.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
nest-asyncio = "^1.6.0"
python-dotenv = "^1.0.1"
kaggle = "^1.6.6"
pretty-jupyter = "^2.0.7"
itables = "^1.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

